# swagger/swagger.yaml
basePath: /v0
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  GwentapiCard:
    description: A card (default view)
    example:
      faction:
        href: Aut aut perspiciatis explicabo neque.
        id: Aliquid recusandae at et sit voluptatum nostrum.
        name: Possimus minus quo et non nulla.
      flavor: Ea quia voluptatibus neque esse.
      href: Qui minus accusantium.
      id: Et quia sunt debitis.
      name: Doloribus est asperiores sequi autem.
      rarity:
        href: Dolores totam dignissimos.
        id: Praesentium deleniti quia voluptate est voluptas.
        name: Eum rerum et.
      rows:
      - Possimus ut et veniam.
      strength: 3.7639757e+18
      subtypes:
      - href: Eveniet est quisquam unde qui eius aut.
        id: Dolorem at numquam.
        name: Libero illo nam nobis ut.
      - href: Nulla nihil officiis sit consequuntur.
        id: Nisi autem dolores.
        name: Magni autem in.
      - href: Animi id distinctio ut.
        id: Explicabo et inventore laudantium ea.
        name: Animi est.
      text: Animi commodi ut qui voluptate voluptas.
      type:
        href: Architecto ut et.
        id: Sit quas harum.
        name: Accusamus et eum omnis ratione.
    properties:
      faction:
        $ref: '#/definitions/GwentapiFaction'
      flavor:
        description: Flavor text of the card
        example: Architecto saepe quod aperiam.
        type: string
      href:
        description: API href for making requests on the card
        example: Vitae magnam ducimus molestiae omnis ratione id.
        type: string
      id:
        description: Unique card ID
        example: Aperiam architecto nobis voluptatem.
        type: string
      name:
        description: Name of the card
        example: Saepe illum voluptas unde.
        type: string
      rarity:
        $ref: '#/definitions/GwentapiRarity'
      rows:
        description: Rows where the card can be played
        example:
        - Qui architecto dignissimos.
        - Nemo autem voluptatem dolorem laudantium eum.
        - Eum voluptatum rerum natus.
        items:
          example: Modi quibusdam et.
          type: string
        type: array
      strength:
        description: Strength of the card
        example: 1.4303367e+18
        format: int64
        type: integer
      subtypes:
        $ref: '#/definitions/GwentapiTypeCollection'
      text:
        description: Text of the card detailing its abilities and how it plays
        example: Est eveniet at sit sit.
        type: string
      type:
        $ref: '#/definitions/GwentapiType'
    required:
    - id
    - href
    - name
    - type
    - faction
    - rarity
    title: 'Mediatype identifier: application/vnd.gwentapi.card+json'
    type: object
  GwentapiCardCollection:
    description: GwentapiCardCollection is the media type for an array of GwentapiCard
      (default view)
    example:
    - faction:
        href: Nam porro mollitia.
        id: Rerum quia ut quia numquam.
        name: Voluptatem cupiditate id voluptatem laborum.
      flavor: Itaque nisi autem quis sed.
      href: Tempora optio explicabo architecto minus.
      id: Vero praesentium ut iusto rem.
      name: Eos quisquam non.
      rarity:
        href: Et dolores quis occaecati.
        id: Ipsa est.
        name: Laborum eveniet quia dicta ut.
      rows:
      - Ratione maiores consequuntur officiis.
      - Odio perspiciatis repudiandae eos hic.
      strength: 8.28949e+18
      subtypes:
      - href: Maiores accusantium ullam veritatis molestiae quis.
        id: Est rerum ab.
        name: Non eius.
      - href: Quaerat atque vel optio.
        id: Excepturi corporis officiis non.
        name: Aliquam ducimus atque sit cumque.
      text: Quae possimus accusantium repellendus voluptates quis velit.
      type:
        href: Inventore hic harum earum optio.
        id: Nam sit consequuntur consequatur quia.
        name: Dolorem laudantium.
    items:
      $ref: '#/definitions/GwentapiCard'
    title: 'Mediatype identifier: application/vnd.gwentapi.card+json; type=collection'
    type: array
  GwentapiFaction:
    description: A card faction (default view)
    example:
      href: Laborum sint ut veniam.
      id: Reiciendis error sequi quo.
      name: Cumque illum tempora aut numquam minima corporis.
    properties:
      href:
        description: API href for making requests on the faction
        example: Ut et non.
        type: string
      id:
        description: Unique faction ID
        example: Veritatis occaecati.
        type: string
      name:
        description: Name of the faction
        example: Quia fugiat culpa.
        type: string
    required:
    - id
    - href
    - name
    title: 'Mediatype identifier: application/vnd.gwentapi.faction+json'
    type: object
  GwentapiFactionCollection:
    description: GwentapiFactionCollection is the media type for an array of GwentapiFaction
      (default view)
    example:
    - href: Id sunt et est sint aut a.
      id: Sit cupiditate debitis.
      name: Aut aut ullam et ut.
    - href: Laborum est enim quidem placeat cum consequatur.
      id: Esse non incidunt dolores laboriosam.
      name: Magni repudiandae.
    - href: Et quo dolorem dolorem eum.
      id: Nesciunt in enim.
      name: Repellendus eum sunt autem ea.
    items:
      $ref: '#/definitions/GwentapiFaction'
    title: 'Mediatype identifier: application/vnd.gwentapi.faction+json; type=collection'
    type: array
...
host: localhost:8080
info:
  contact:
    email: gwentapi@gmail.com
    name: Tristan S.
    url: https://twitter.com/GwentAPI
  description: A consumption-only API of all Gwent data
  license:
    name: 'The textual information presented through this API about GWENT: The Witcher
      Card Game is copyrighted by CD Projekt RED'
  title: Gwent API
  version: "0.0"
paths:
...
  /cards:
    get:
      description: Return all cards.
      operationId: card#list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiCardCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: list card
      tags:
      - card
  /cards/{cardID}:
    get:
      description: Return card with given id.
      operationId: card#show
      parameters:
      - description: Card ID
        in: path
        name: cardID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiCard'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: show card
      tags:
      - card
  /cards/factions/{factionID}:
    get:
      description: Return all cards with given faction id.
      operationId: card#cardFaction
      parameters:
      - description: Faction ID
        in: path
        name: factionID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiCardCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: cardFaction card
      tags:
      - card
  /cards/leaders:
    get:
      description: Return all leader cards.
      operationId: card#cardLeader
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiCardCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: cardLeader card
      tags:
      - card
  /cards/rarities/{rarityID}:
    get:
      description: Return all cards with given rarity id.
      operationId: card#cardRarity
      parameters:
      - description: Rarity ID
        in: path
        name: rarityID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiCardCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: cardRarity card
      tags:
      - card
  /factions:
    get:
      description: Return all factions.
      operationId: faction#list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiFactionCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: list faction
      tags:
      - faction
  /factions/{factionID}:
    get:
      description: Return faction with given id.
      operationId: faction#show
      parameters:
      - description: Faction ID
        in: path
        name: factionID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiFaction'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: show faction
      tags:
      - faction
  /glyphs:
    get:
      description: Return all glyphs.
      operationId: glyph#list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiGlyphCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: list glyph
      tags:
      - glyph
  /glyphs/{glyphID}:
    get:
      description: Return glyph with given id.
      operationId: glyph#show
      parameters:
      - description: Glyph ID
        in: path
        name: glyphID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiGlyph'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: show glyph
      tags:
      - glyph
  /patches:
    get:
      description: Return all patches.
      operationId: patch#list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiPatchCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: list patch
      tags:
      - patch
  /patches/{patchID}:
    get:
      description: Return patch with given id.
      operationId: patch#show
      parameters:
      - description: Patch ID
        in: path
        name: patchID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiPatch'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: show patch
      tags:
      - patch
  /patches/latest:
    get:
      description: Return latest patch.
      operationId: patch#latest
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiPatch'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: latest patch
      tags:
      - patch
  /rarities:
    get:
      description: Return all rarities.
      operationId: rarity#list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiRarityCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: list rarity
      tags:
      - rarity
  /rarities/{rarityID}:
    get:
      description: Return rarity with given id.
      operationId: rarity#show
      parameters:
      - description: Rarity ID
        in: path
        name: rarityID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiRarity'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: show rarity
      tags:
      - rarity
  /types:
    get:
      description: Return all card types.
      operationId: type#list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiTypeCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: list type
      tags:
      - type
  /types/{typeID}:
    get:
      description: Return card type with given id.
      operationId: type#show
      parameters:
      - description: Card type ID
        in: path
        name: typeID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GwentapiType'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: show type
      tags:
      - type
produces:
- application/json
...
schemes:
- https
swagger: "2.0"